+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.1] - 2025-01-28
 ### Fixed
+- Vec2 argument type
 - Vec3 argument type
 
 ## [2.0.0+beta.2] - 2024-10-04